fix: 收敛 core_family 静态分析修复与 agent 策略
[Summary]
完成 core_family 队列中的一批源码修复，并将 clang-tidy workflow 收敛为更纯粹的 agent policy。

[Changed & Refactored]
- 调整 transport、core、bridge 与 adapters 相关实现，消除一批 core_family clang-tidy 告警
- 将 time_tracer 与 log_generator 的 clang-tidy workflow 压缩为面向 agent 的策略文件
- 统一单任务处理顺序为 patch -> fix --dry-run -> suggest -> step -> tidy-batch

[Fixed]
- 修复多处局部常量命名、冗余转换与重复解析骨架带来的静态分析问题
- 修复 workflow 中夹杂大量 CLI 帮助手册内容、导致 agent 决策噪声过高的问题

[Verification]
- python tools/run.py verify --app tracer_core_shell --build-dir build_fast --concise --scope task

Release-Version: v0.8.0

+++ libs/tracer_core/src/infrastructure/crypto/internal/file_crypto_kdf.cpp
@@ -94,8 +94,7 @@ auto ResolvePwhashLimits(FileCryptoSecurityLevel security_level)
       return {
           .ops_limit =
               static_cast<unsigned long long>(crypto_pwhash_OPSLIMIT_MAX),
-          .kMemLimitBytes =
-              static_cast<unsigned long long>(crypto_pwhash_MEMLIMIT_MAX),
+          .kMemLimitBytes = crypto_pwhash_MEMLIMIT_MAX,
       };
     case FileCryptoSecurityLevel::kModerate:
       return {
